+++ installer/configurator/go.mod
@@ -1,0 +1,7 @@
+module github.com/DataDog/nginx-datadog/installer
+
+go 1.21.0
+
+require github.com/google/go-github v17.0.0+incompatible
+
+require github.com/google/go-querystring v1.1.0 // indirect
