--- rum-injection-installer/installer/go.mod
@@ -1,3 +1,0 @@
-module github.com/DataDog/rum-injection-installer/installer
-
-go 1.21.0
